refactor: Header에 있는 Navigator들 추상화 및 폴더 정리
