+++ src/components/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Header from './common/Header';
+import Header from './Header';
 
 const Layout = () => {
   return (
